Move train_logreg to annotation.py and compute features in there
