+++ tools/sense_studio/annotation.py
@@ -10,10 +10,13 @@ from flask import render_template
 from flask import request
 from flask import send_from_directory
 from flask import url_for
+from joblib import dump
 from joblib import load
 from natsort import natsorted
 from natsort import ns
+from sklearn.linear_model import LogisticRegression
 
+from sense import SPLITS
 from sense.finetuning import compute_frames_and_features
 from tools import directories
 from tools.sense_studio import project_utils
@@ -169,20 +172,7 @@ def submit_annotation():
 
     # Automatic re-training of the logistic regression model
     if utils.get_project_setting(path, 'assisted_tagging'):
-        inference_engine, model_config = utils.load_feature_extractor(path)
-        videos_dir = directories.get_videos_dir(path, split, label)
-        frames_dir = directories.get_frames_dir(path, split, label)
-        features_dir = directories.get_features_dir(path, split, model_config, label=label)
-
-        # Compute the respective frames and features
-        compute_frames_and_features(inference_engine=inference_engine,
-                                    project_path=path,
-                                    videos_dir=videos_dir,
-                                    frames_dir=frames_dir,
-                                    features_dir=features_dir)
-
-        # Re-train the logistic regression model
-        utils.train_logreg(path=path, label=label)
+        train_logreg(path=path, label=label)
 
     if next_frame_idx >= len(os.listdir(frames_dir)):
         return redirect(url_for('.show_video_list', project=project, split=split, label=label))
@@ -198,3 +188,63 @@ def download_file(project, split, label, video_name, img_file):
     dataset_path = project_utils.lookup_project_path(project)
     img_dir = os.path.join(directories.get_frames_dir(dataset_path, split, label), video_name)
     return send_from_directory(img_dir, img_file, as_attachment=True)
+
+
+def train_logreg(path, label):
+    """
+    (Re-)Train a logistic regression model on all annotations that have been submitted so far.
+    """
+    inference_engine, model_config = project_utils.load_feature_extractor(path)
+
+    logreg_dir = directories.get_logreg_dir(path, model_config, label)
+    logreg_path = os.path.join(logreg_dir, 'logreg.joblib')
+    project_config = project_utils.load_project_config(path)
+    class_tags = project_config['classes'][label]
+
+    all_features = []
+    all_annotations = []
+
+    for split in SPLITS:
+        videos_dir = directories.get_videos_dir(path, split, label)
+        frames_dir = directories.get_frames_dir(path, split, label)
+        features_dir = directories.get_features_dir(path, split, model_config, label=label)
+        tags_dir = directories.get_tags_dir(path, split, label)
+
+        if not os.path.exists(tags_dir):
+            continue
+
+        # Compute the respective frames and features
+        compute_frames_and_features(inference_engine=inference_engine,
+                                    project_path=path,
+                                    videos_dir=videos_dir,
+                                    frames_dir=frames_dir,
+                                    features_dir=features_dir)
+
+        video_tag_files = os.listdir(tags_dir)
+
+        for video_tag_file in video_tag_files:
+            feature_file = os.path.join(features_dir, video_tag_file.replace('.json', '.npy'))
+            annotation_file = os.path.join(tags_dir, video_tag_file)
+
+            features = np.load(feature_file)
+            for f in features:
+                all_features.append(f.mean(axis=(1, 2)))
+
+            with open(annotation_file, 'r') as f:
+                all_annotations.extend(json.load(f)['time_annotation'])
+
+    # Reset tags that have been removed from the class to 'background'
+    all_annotations = [tag_idx if tag_idx in class_tags else 0 for tag_idx in all_annotations]
+
+    # Use low class weight for background and higher weight for all present tags
+    annotated_tags = set(all_annotations)
+    class_weight = {tag: 2 for tag in annotated_tags}
+    class_weight[0] = 0.5
+
+    all_features = np.array(all_features)
+    all_annotations = np.array(all_annotations)
+
+    if len(annotated_tags) > 1:
+        logreg = LogisticRegression(C=0.1, class_weight=class_weight)
+        logreg.fit(all_features, all_annotations)
+        dump(logreg, logreg_path)
