+++ tools/sense_studio/sense_studio.py
@@ -20,12 +20,11 @@ from flask import request
 from flask import url_for
 
 from sense import SPLITS
-from sense.finetuning import compute_frames_and_features
 from tools import directories
 from tools.sense_studio import project_utils
 from tools.sense_studio import socketio
-from tools.sense_studio import utils
 from tools.sense_studio.annotation import annotation_bp
+from tools.sense_studio.annotation import train_logreg
 from tools.sense_studio.testing import testing_bp
 from tools.sense_studio.training import training_bp
 from tools.sense_studio.video_recording import video_recording_bp
@@ -247,23 +246,10 @@ def toggle_project_setting():
 
     # Update logreg model if assisted tagging was just enabled
     if setting == 'assisted_tagging' and new_status:
-        split = data['split']
         label = data['label']
-        inference_engine, model_config = utils.load_feature_extractor(path)
-
-        videos_dir = directories.get_videos_dir(path, split, label)
-        frames_dir = directories.get_frames_dir(path, split, label)
-        features_dir = directories.get_features_dir(path, split, model_config, label=label)
-
-        # Compute the respective frames and features
-        compute_frames_and_features(inference_engine=inference_engine,
-                                    project_path=path,
-                                    videos_dir=videos_dir,
-                                    frames_dir=frames_dir,
-                                    features_dir=features_dir)
 
         # Re-train the logistic regression model
-        utils.train_logreg(path=path, label=label)
+        train_logreg(path=path, label=label)
 
     return jsonify(setting_status=new_status)
